Struct trillium_websockets::WebSocket

pub struct WebSocket<Handler> { /* fields omitted */ }

The trillium handler. See crate-level docs for example usage.

Implementations

impl<Handler, Fut> WebSocket<Handler> where
Handler: Fn(WebSocketConn) -> Fut + Sync + Send + 'static,
Fut: Future<Output = ()> + Send + 'static,

pub fn new(handler: Handler) -> Self

Build a new WebSocket with an async handler function that receives a WebSocketConn

pub fn with_protocols(self, protocols: &[&str]) -> Self

protocols is a sequence of known protocols. On successful handshake, the returned response headers contain the first protocol in this list which the server also knows.