Struct trillium_websockets::WebSocketConn

pub struct WebSocketConn { /* private fields */ }

A struct that represents an specific websocket connection.

This can be thought of as a combination of a [async_tungstenite::WebSocketStream] and a [trillium::Conn], as it contains a combination of their fields and associated functions.

The WebSocketConn implements Stream<Item=Result<Message, Error>>, and can be polled with StreamExt::next

Implementations

impl WebSocketConn

pub async fn send_string(&mut self, string: String) -> Result<()>

send a Message::Text variant

pub async fn send_bytes(&mut self, bin: Vec<u8>) -> Result<()>

send a Message::Binary variant

pub async fn send(&mut self, message: Message) -> Result<()>

Sends a Message to the client

pub fn stopper(&self) -> Stopper

retrieve a clone of the server’s [Stopper]

pub async fn close(&mut self) -> Result<()>

close the websocket connection gracefully

pub fn headers(&self) -> &Headers

retrieve the request headers for this conn

pub fn peer_ip(&self) -> Option<IpAddr>

retrieves the peer ip for this conn, if available

pub fn path(&self) -> &str

retrieves the path part of the request url, up to and excluding any query component

pub fn querystring(&self) -> &str

Retrieves the query component of the path, excluding ?. Returns an empty string if there is no query component.

pub fn method(&self) -> Method

retrieve the request method for this conn

pub fn state<T: 'static>(&self) -> Option<&T>

retrieve state from the state set that has been accumulated by trillium handlers run on the [trillium::Conn] before it became a websocket. see [trillium::Conn::state] for more information

pub fn state_mut<T: 'static>(&mut self) -> Option<&mut T>

retrieve a mutable borrow of the state from the state set

pub fn set_state<T: Send + Sync + 'static>(&mut self, val: T)

set state on this connection

pub fn take_state<T: 'static>(&mut self) -> Option<T>

take some type T out of the state set that has been accumulated by trillium handlers run on the [trillium::Conn] before it became a websocket. see [trillium::Conn::take_state] for more information

pub fn take_inbound_stream( &mut self ) -> Option<impl Stream<Item = Result<Message, Error>>>

take the inbound Message stream from this conn

pub fn inbound_stream( &mut self ) -> Option<impl Stream<Item = Result<Message, Error>> + '_>

borrow the inbound Message stream from this conn