Struct trillium_websockets::WebSocket

pub struct WebSocket<H> { /* private fields */ }

The trillium handler. See crate-level docs for example usage.

Implementations

impl<H> WebSocket<H> where
H: WebSocketHandler,

pub fn new(handler: H) -> Self

Build a new WebSocket with an async handler function that receives a WebSocketConn

pub fn with_protocols(self, protocols: &[&str]) -> Self

protocols is a sequence of known protocols. On successful handshake, the returned response headers contain the first protocol in this list which the server also knows.

pub fn with_protocol_config(self, config: WebSocketConfig) -> Self

configure the websocket protocol